ALLD_alloc was calling ERRD_bugcheck() when it doesn't have to call it due to a bad implicit cast. No call to isc_dsql_allocate_statement would succeed due to this small issue.

+++ dsql/alld.c
@@ -271,7 +271,7 @@ memset (block, 0, BLOCKS_TO_SIZE (needed_blocks));
 block->blk_type = type;
 block->blk_pool_id_mod = (UCHAR)pool->plb_pool_id;
 
-if (needed_blocks > MAX_USHORT)
+if (needed_blocks > (ULONG) MAX_USHORT)
     BUGCHECK ("ALLD_alloc: too greedy");
 
 block->blk_length = (USHORT)needed_blocks;
